All tests passing

+++ src/position2d.cpp
@@ -1,4 +1,6 @@
 #include "position2d.h"
+#include "puzzleboard.h"
+
 
 Position2D::Position2D(int x, int y):
     X(x),Y(y)
@@ -8,7 +10,8 @@ Position2D::Position2D(int x, int y):
 
 bool Position2D::positionExceedsBounds(const Dimension2D &d)
 {
-    if (X >= d.horizontalSize || Y >= d.verticalSize)
+    if (X >= (int)d.horizontalSize || Y >= (int)d.verticalSize)
         return true;
     return false;
 }
+
